--- apps/github-receiver/src/core/interfaces/process-result.interface.ts
@@ -1,5 +1,0 @@
-export interface ProcessResult {
-  success: boolean;
-  message: string;
-  error?: Error;
-}
