--- apps/github-receiver/src/core/processors/index.ts
@@ -1,1 +1,0 @@
-export { BaseEventProcessor, PullRequestEventProcessor, PushEventProcessor, IssueEventProcessor } from './github-event.processor';
